+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/member/presentation/controller/MemberController.java
@@ -2,7 +2,7 @@ package com.nextdoor.nextdoor.domain.member.presentation.controller;
 
 import com.nextdoor.nextdoor.domain.member.presentation.dto.request.MemberExtraInfoSaveRequestDto;
 import com.nextdoor.nextdoor.domain.member.presentation.dto.response.MemberResponseDto;
-import com.nextdoor.nextdoor.domain.member.application.MemberService;
+import com.nextdoor.nextdoor.domain.member.application.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
